[REF] shopinvader_variant_alias: split models into one file per model
