+++ shopinvader_variant_alias/models/__init__.py
@@ -1,3 +1,5 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from . import models
+from . import product_attribute_value
+from . import product_attribute_template_value
+from . import shopinvader_variant
